@@ -1,18 +1,16 @@
-'use strict';
+import path from 'path';
 
-const path = require('path');
+import fs from 'fs-extra';
+import {promisify} from 'util';
+// import check from 'check-types-2';
+import bump from 'bump-regex';
 
-const fs = require('fs-extra');
-const {promisify} = require('util');
-// const check = require('check-types-2');
-const bump = require('bump-regex');
+import ui from '../ui.js';
+import * as git from '../git/index.js';
 
 const asyncBump = promisify(bump);
 
-const ui = require('../ui');
-const git = require('../git');
-
-module.exports = {
+export default {
   description: 'Release',
   options: [
     {
