@@ -1,10 +1,8 @@
-'use strict';
+import {execa}      from 'execa';
+import check        from 'check-types-2';
+import {processenv} from 'processenv';
 
-const execa        = require('execa');
-const check        = require('check-types-2');
-const {processenv} = require('processenv');
-
-module.exports = async(cmd, options = {}) => {
+export default async(cmd, options = {}) => {
   check.assert.string(cmd, 'cmd is not a string');
 
   const cwd = options.cwd || process.cwd();
