@@ -1,6 +1,4 @@
-'use strict';
-
-const log4js     = require('log4js');
+import log4js from 'log4js';
 
 const logger = log4js.getLogger();
 
@@ -20,9 +18,9 @@ Object.assign(logger, {
   taskError: (name, err) => {
     logger.info(`> ${name} failed`, err);
   },
-  taskSuccess: (name, err) => {
+  taskSuccess: name => {
     logger.info(`> ${name} succeeded`);
-  }
+  },
 });
 
-module.exports = logger;
+export default logger;
